Add --log Parameter To Control Logging Level
Add --log parameter that allows you to pick the log level used.  If none
is set it uses the library default of warning.  Uses the standard values
of DEBUG, INFO, WARNING, ERROR or CRITICAL.

The only problem with this is that it makes things silent all the time.
I will fix that soon.